Fix cron job stale hint text, deduplicate approxIntervalMs, remove unused dayMs
- Fix hard-coded stale warning text: Now dynamically computed based on job interval (shows '>36h' for daily jobs, '>6h' for 4h jobs, etc.)
- Extract approxIntervalMs to shared.ts to eliminate duplication between register-storage-and-stats.ts and cmd-verify.ts
- Remove unused dayMs variable in register-storage-and-stats.ts
